ping: move node info query functions to a sepate file
The node info query stuff is pretty distinct so move it to it's own file.
